Rename module to structured.generic
