+++ tests/v4/test_gridadapter.py
@@ -4,7 +4,7 @@ import numpy as np
 import pytest
 from numpy.testing import assert_allclose
 
-from parcels._datasets.structured.grid_datasets import N, T, datasets
+from parcels._datasets.structured.generic import N, T, datasets
 from parcels.grid import Grid as OldGrid
 from parcels.tools.converters import TimeConverter
 from parcels.v4.grid import Grid as NewGrid
